Refactor fbx writing
